feat: prohibit kick dealer, little fix
